shortcut-editor: use states to handle headerbar mode
Instead of manually handling every button in the headerbar
individually, using states to set it is clearer and makes
the code easier to read and understand.

This patch adds the headerbar mode handling code, and updates
the current code to use it rather than by updating each individual
button manually.

https://bugzilla.gnome.org/show_bug.cgi?id=769314