display/config: Add support for setting new configuration
Add support, to both the internal API and to the console client, to set
a new configuration.

For example, if there is a system with two monitors, LVDS1 and LVDS2,
both with 1024x768 preferred resolution, to configure them to be placed
vertically, one can now do:

./cc-display-test set -L -p -M LVDS1 -L -y 768 -M LVDS2

See ./cc-display-test --help for more details.